Fix IAS header file inclusion
Use of include_directories() is causing the IAS_INCLUDES path to be
included in local library builds.  This means that headers that have
been installed in IAS_INCLUDES are overriding the local headers, and
these headers may be from different branches.
Switching to use of target_include_directories() for including the
IAS_INCLUDES headers resolves this.