LDEV-820 Convert ESPA raw format to GeoTIFF using IAS library
Use the IAS GeoTIFF library support to write the output GeoTIFF image
when converting from an ESPA raw image.  The IAS library is used
when the datum is WGS84.  For other datums, it still falls back to
using GDAL since IAS only supports WGS84.  The IAS library is used
to make sure the Level 2 GeoTIFFs match the Level 1 files.